Refactor More Casper Functions Into Beacon-Chain/Core (#890)
