--- beacon-chain/casper/BUILD.bazel
@@ -1,40 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = [
-        "sharding.go",
-        "state_transition.go",
-    ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/casper",
-    visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//beacon-chain/core/incentives:go_default_library",
-        "//beacon-chain/core/validators:go_default_library",
-        "//beacon-chain/utils:go_default_library",
-        "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bitutil:go_default_library",
-        "//shared/params:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    size = "large",
-    timeout = "short",
-    srcs = [
-        "sharding_test.go",
-        "state_transition_test.go",
-    ],
-    embed = [":go_default_library"],
-    race = "off",  # TODO(#604): fix issues with tests failing with race on.
-    deps = [
-        "//beacon-chain/core/validators:go_default_library",
-        "//beacon-chain/utils:go_default_library",
-        "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bytes:go_default_library",
-        "//shared/params:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
-    ],
-)
