+++ beacon-chain/core/incentives/incentives.go
@@ -5,10 +5,41 @@
 package incentives
 
 import (
+	"github.com/prysmaticlabs/prysm/beacon-chain/utils"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
+	"github.com/prysmaticlabs/prysm/shared/bitutil"
 	"github.com/prysmaticlabs/prysm/shared/mathutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 )
+
+// TallyVoteBalances calculates all the votes behind a block and
+// then rewards validators for their participation in voting for that block.
+func TallyVoteBalances(
+	blockHash [32]byte,
+	blockVoteCache utils.BlockVoteCache,
+	validators []*pb.ValidatorRecord,
+	activeValidatorIndices []uint32,
+	totalActiveValidatorDeposit uint64,
+	timeSinceFinality uint64,
+) (uint64, []*pb.ValidatorRecord) {
+	blockVote, ok := blockVoteCache[blockHash]
+	if !ok {
+		return 0, validators
+	}
+
+	blockVoteBalance := blockVote.VoteTotalDeposit
+	voterIndices := blockVote.VoterIndices
+	validators = CalculateRewards(
+		voterIndices,
+		activeValidatorIndices,
+		validators,
+		totalActiveValidatorDeposit,
+		blockVoteBalance,
+		timeSinceFinality,
+	)
+
+	return blockVoteBalance, validators
+}
 
 // CalculateRewards adjusts validators balances by applying rewards or penalties
 // based on FFG incentive structure.
@@ -35,7 +66,7 @@ func CalculateRewards(
 				if voterIndex == validatorIndex {
 					voted = true
 					balance := validators[validatorIndex].GetBalance()
-					newBalance := calculateBalance(balance, rewardQuotient, totalParticipatedDeposit, totalActiveValidatorDeposit)
+					newBalance := int64(balance) + int64(balance/rewardQuotient)*(2*int64(totalParticipatedDeposit)-int64(totalActiveValidatorDeposit))/int64(totalActiveValidatorDeposit)
 					validators[validatorIndex].Balance = uint64(newBalance)
 					break
 				}
@@ -71,6 +102,35 @@ func CalculateRewards(
 	return validators
 }
 
+// ApplyCrosslinkRewardsAndPenalties applies the appropriate rewards and
+// penalties according to the attestation for a shard.
+func ApplyCrosslinkRewardsAndPenalties(
+	crosslinkRecords []*pb.CrosslinkRecord,
+	slot uint64,
+	attesterIndices []uint32,
+	attestation *pb.AggregatedAttestation,
+	validators []*pb.ValidatorRecord,
+	totalActiveValidatorDeposit uint64,
+	totalBalance uint64,
+	voteBalance uint64,
+) error {
+	rewardQuotient := RewardQuotient(totalActiveValidatorDeposit)
+	for _, attesterIndex := range attesterIndices {
+		timeSinceLastConfirmation := slot - crosslinkRecords[attestation.Shard].GetSlot()
+
+		checkBit, err := bitutil.CheckBit(attestation.AttesterBitfield, int(attesterIndex))
+		if err != nil {
+			return err
+		}
+		if checkBit {
+			RewardValidatorCrosslink(totalBalance, voteBalance, rewardQuotient, validators[attesterIndex])
+		} else {
+			PenaliseValidatorCrosslink(timeSinceLastConfirmation, rewardQuotient, validators[attesterIndex])
+		}
+	}
+	return nil
+}
+
 // RewardQuotient returns the reward quotient for validators which will be used to
 // reward validators for voting on blocks, or penalise them for being offline.
 func RewardQuotient(totalActiveValidatorDeposit uint64) uint64 {
@@ -93,55 +153,27 @@ func QuadraticPenalty(numberOfSlots uint64) uint64 {
 	return slotFactor / penaltyQuotient
 }
 
-// RewardValidatorCrosslink applies rewards to validators part of a shard
-// committee for voting on a shard.
+// RewardValidatorCrosslink applies rewards to validators part of a shard committee for voting on a shard.
+// TODO(#538): Change this to big.Int as tests using 64 bit integers fail due to integer overflow.
 func RewardValidatorCrosslink(
 	totalDeposit uint64,
 	participatedDeposits uint64,
 	rewardQuotient uint64,
 	validator *pb.ValidatorRecord,
-) *pb.ValidatorRecord {
-	balance := calculateBalance(validator.Balance, rewardQuotient, participatedDeposits, totalDeposit)
-	return &pb.ValidatorRecord{
-		Pubkey:            validator.Pubkey,
-		WithdrawalShard:   validator.WithdrawalShard,
-		WithdrawalAddress: validator.WithdrawalAddress,
-		RandaoCommitment:  validator.RandaoCommitment,
-		Balance:           balance,
-		Status:            validator.Status,
-		ExitSlot:          validator.ExitSlot,
-	}
+) {
+	currentBalance := int64(validator.Balance)
+	currentBalance += int64(currentBalance) / int64(rewardQuotient) * (2*int64(participatedDeposits) - int64(totalDeposit)) / int64(totalDeposit)
+	validator.Balance = uint64(currentBalance)
 }
 
-// PenaliseValidatorCrosslink applies penalties to validators part of a shard committee
-// for not voting on a shard.
+// PenaliseValidatorCrosslink applies penalties to validators part of a shard committee for not voting on a shard.
 func PenaliseValidatorCrosslink(
 	timeSinceLastConfirmation uint64,
 	rewardQuotient uint64,
 	validator *pb.ValidatorRecord,
-) *pb.ValidatorRecord {
+) {
+	newBalance := validator.Balance
 	quadraticQuotient := QuadraticPenaltyQuotient()
-	balance := validator.Balance
-	balance -= balance/rewardQuotient + balance*timeSinceLastConfirmation/quadraticQuotient
-	return &pb.ValidatorRecord{
-		Pubkey:            validator.Pubkey,
-		WithdrawalShard:   validator.WithdrawalShard,
-		WithdrawalAddress: validator.WithdrawalAddress,
-		RandaoCommitment:  validator.RandaoCommitment,
-		Balance:           balance,
-		Status:            validator.Status,
-		ExitSlot:          validator.ExitSlot,
-	}
-}
-
-// calculateBalance applies the Casper FFG reward calculation based on reward quotients
-// and total deposits from validators.
-func calculateBalance(
-	balance uint64,
-	rewardQuotient uint64,
-	totalParticipatedDeposit uint64,
-	totalActiveValidatorDeposit uint64,
-) uint64 {
-	participationNumerator := 2*int64(totalParticipatedDeposit) - int64(totalActiveValidatorDeposit)
-	return uint64(int64(balance) + int64(balance/rewardQuotient)*participationNumerator/int64(totalActiveValidatorDeposit))
+	newBalance -= newBalance/rewardQuotient + newBalance*timeSinceLastConfirmation/quadraticQuotient
+	validator.Balance = newBalance
 }
