+++ beacon-chain/core/state/BUILD.bazel
@@ -2,25 +2,23 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["incentives.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/incentives",
+    srcs = ["processing.go"],
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/core/state",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//beacon-chain/utils:go_default_library",
+        "//beacon-chain/core/validators:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/bitutil:go_default_library",
-        "//shared/mathutil:go_default_library",
         "//shared/params:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["incentives_test.go"],
+    srcs = ["processing_test.go"],
     embed = [":go_default_library"],
     deps = [
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared/mathutil:go_default_library",
+        "//shared/bytes:go_default_library",
         "//shared/params:go_default_library",
     ],
 )
