+++ beacon-chain/types/BUILD.bazel
@@ -11,7 +11,8 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/types",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//beacon-chain/casper:go_default_library",
+        "//beacon-chain/core/incentives:go_default_library",
+        "//beacon-chain/core/state:go_default_library",
         "//beacon-chain/core/validators:go_default_library",
         "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
@@ -35,9 +36,9 @@ go_test(
         "crystallized_state_test.go",
     ],
     embed = [":go_default_library"],
-    race = "off",  # TODO(#604): fix issues with tests failing with race on.
+    race = "on",
     deps = [
-        "//beacon-chain/casper:go_default_library",
+        "//beacon-chain/core/validators:go_default_library",
         "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/hashutil:go_default_library",
